Ajustes pré interface

+++ Dropbox/00 - IME/5s/MAC0323 - Algoritmos e Estruturas de Dados II/EPs/ep1/ep1.cpp
@@ -61,7 +61,7 @@ int main() {
 
 
 string minuscula(string s) {
-    for (int i = 0; i < s.length(); i++)
+    for (int i = 0; (unsigned)i < s.length(); i++)
         s[i] = tolower(s[i]);
     return s;
 }
